#BUG01 - Testando variações de JSON/Dictionaries
